scripts/openstack-exporter-app/requirements.txt
prometheus_client
requests
keystoneauth1
pyyaml
openstacksdk
